feat(issue-details): Show static replay when error is not within the replay
